requirements.txt
Flask==2.2.2
python-dotenv==1.0.0
requests==2.28.2
beautifulsoup4==4.12.2
selenium==4.9.0
appopener==1.7
